Derive the snapshot through the shared schema walk and platform registry
